코드 분리(with ai)
